move general state adjustment routines to hiccup_utilities
